chore: move functions to utils
